// Izvorni kod/shared/enums/SurveyEnums.ts
export enum FirstQuestion {
  NITKO = 1,
  POJEDINCI = 2,
  MNOGI = 3,
  VECINA = 5,
}

export enum SecondQuestion {
  NITKO = 3,
  POJEDINCI = 4,
  MNOGI = 5,
  VECINA = 6,
}

export enum ThirdQuestion {
  SE_NISU_TRESLI = 1,
  SU_SE_TRESLI_SLABO = 3,
  SU_SE_TRESLI_UMJERENO = 4,
  SU_SE_TRESLI_JAKO = 5,
  SU_SE_POMICALI = 6,
  SU_PADALI = 7,
}

export enum FourthQuestion {
  SE_NISU_TRESLI = 2,
  SU_SE_TRESLI = 3,
  SU_ZVECALI = 4,
  SU_UDARALI = 5,
  SU_SE_OTVARALI_I_ZATVARALI = 6,
  SU_SE_RAZBILI = 7,
}

export enum FifthQuestion {
  SE_NISU_NJIHALI = 2,
  SU_SE_SLABO_NJIHALI = 3,
  SU_SE_UMJERENO_NJIHALI = 4,
  SU_SE_JAKO_NJIHALI = 5,
  SU_PADALI = 6,
}

export enum SixthQuestion {
  NIJE_BUDIO = 3,
  BUDIO_JE_POJEDINCE = 4,
  BUDIO_JE_MNOGE = 5,
  BUDIO_JE_VECINU = 6,
}

export enum SeventhQuestion {
  NE = 4,
  POJEDINCI = 5,
  MNOGI = 6,
  VECINA = 7,
}

export enum EighthQuestion {
  NISU = 5,
  ZVONILA_MALA_ZVONA = 6,
  ZVONILA_VELIKA_ZVONA = 7,
}

export enum NinthQuestion {
  NITI_JEDNA = 7,
  PONEKA = 8,
  MNOGE = 10,
  VECINA = 11,
  SVE = 12,
}

export enum TenthQuestion {
  JE = 8,
  NIJE = 7,
}